Add files via upload
This encoder automates the encoding process used originally by Mati Aharoni. The original encoder was found in the book "Hacking: The Art of Exploitation" written by John Ericson. I just changed the source a bit in order to fit our needs. The character set mentioned in the source code is for this exploit written by MUTS in the link below:
https://www.exploit-db.com/exploits/5342/